♻️refactor: 테스트 수정 & 적절한 위치로 파일 이동
