+++ src/main/java/util/stylesheetUtils.java
@@ -1,5 +1,7 @@
 package util;
 
+import model.ContentType;
+
 public class StylesheetUtils {
     public static String getContentType(String url) {
         return ContentType.of(url).getContentType();
